// src/api/model/truck.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * OLHSO API
 * The OLHSO API description
 * OpenAPI spec version: 0.1
 */
import type { TruckArea } from './truckArea';
import type { TruckCoordinates } from './truckCoordinates';
import type { DrivingStatus } from './drivingStatus';
import type { TruckModalAssignment } from './truckModalAssignment';
import type { TruckSlotArray } from './truckSlotArray';
import type { TruckWok } from './truckWok';
import type { VehicleStatus } from './vehicleStatus';

export interface Truck {
  areas?: TruckArea[];
  coordinates?: TruckCoordinates;
  createdAt: string;
  distance?: number;
  drivingStatus: DrivingStatus;
  id: string;
  licensePlateNumber: string;
  modalAssignment?: TruckModalAssignment[];
  name: string;
  note?: string;
  slotArray?: TruckSlotArray;
  updatedAt: string;
  usedWoks?: TruckWok;
  vehicleStatus: VehicleStatus;
}